refactor(trusty-search): split oversized test files under 500 SLOC cap (#607)
Split two monolithic test files that exceeded the 500-SLOC hard cap into
focused per-concept submodules, each well under the cap:

core/indexer/tests.rs (3426 total lines, ~2472 SLOC) → tests/:
  mod.rs (helpers + imports), persistence.rs, search_basic.rs,
  search_kg.rs, search_ranking.rs, search_mode.rs, entity.rs,
  grep.rs, batch.rs, paths.rs

service/reindex/tests.rs (1674 total lines, ~1113 SLOC) → tests/:
  mod.rs (helpers + imports), walk_filter.rs, context.rs,
  stages.rs, semaphore.rs, corpus.rs

All 867 passing tests remain passing; the one pre-existing failure
(test_kg_results_survive_top_k_truncation, flaky BM25 race) is
unchanged. Stale allowlist entries for the deleted files removed via
check_line_cap.sh --update.
